% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multe.R
\name{multe}
\alias{multe}
\title{Contamination Bias Diagnostics for Multiple Treatments}
\usage{
multe(r, treatment_name, cluster = NULL, tol = 1e-07)
}
\arguments{
\item{r}{Fitted model, output of the \code{lm} function.}

\item{treatment_name}{name of treatment variable}

\item{cluster}{Factor variable that defines clusters. If \code{NULL} (or not
supplied), the command computes heteroscedasticity-robust standard
errors, rather than cluster-robust standard errors.}

\item{tol}{Numerical tolerance for computing LM test statistic}
}
\value{
Returns a list with the following components: TODO
}
\description{
Contamination Bias Diagnostics for Multiple Treatments
}
